+++ Sources/SynchronousData.swift
@@ -3,21 +3,30 @@
 import Foundation
 
 extension Utility {
-    
-    /// Creates a blocking URLSession data task.
+    /// Creates and executes a non-blocking URLSession data task.
+    /// - Parameter request: the `URLRequest` for the session.
+    /// - Parameter completion: A closure to execute asynchronously when the data task completes.
+    /// - Returns: The retrieved session data.
+    public static func requestAsynchronousData(_ request: URLRequest, completion: @escaping (Result<Data, Error>) -> Void) -> Void {
+        let task = URLSession.shared.dataTask(with: request) { data, _, error -> Void in
+            completion(Result(data, error))
+        }
+        task.resume()
+    }
+
+    /// Creates and executes a blocking URLSession data task.
     /// - Parameter request: the `URLRequest` for the session.
     /// - Throws: Any returned error on `Result.get`.
     /// - Returns: The retrieved session data.
     public static func requestSynchronousData(_ request: URLRequest) throws -> Data {
         let semaphore = DispatchSemaphore(value: 0)
-        var result: Result<Data, Error> = Result(Data(), nil)
-        let task = URLSession.shared.dataTask(with: request) { data, _, error -> Void in
-            result = Result(data, error)
-            semaphore.signal()
+        var received: Result<Data, Error> = Result(Data(), nil)
+        let completion = { (result: Result<Data, Error>) -> Void in
+            received = result
+            _ = semaphore.signal()
         }
-        task.resume()
+        requestAsynchronousData(request, completion: completion)
         let _ = semaphore.wait(timeout: DispatchTime.distantFuture)
-        return try result.get()
+        return try received.get()
     }
-
 }